requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
contourpy==1.3.1
cycler==0.12.1
eval_type_backport==0.2.2
fonttools==4.55.7
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jsonpath-python==1.0.6
kiwisolver==1.4.8
matplotlib==3.10.0
mistralai==1.5.0
mypy-extensions==1.0.0
numpy==2.2.2
packaging==24.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sniffio==1.3.1
typing-inspect==0.9.0
typing_extensions==4.12.2
